requirements.txt
# Backend dependencies
fastapi>=0.100.0
uvicorn>=0.23.0
gradio==3.50.2
pandas>=2.1.0
numpy>=1.26.0
matplotlib>=3.8.0
seaborn>=0.13.0
plotly>=5.18.0
requests>=2.31.0
scikit-learn>=1.3.0
python-multipart>=0.0.6
pydantic>=2.0.0
dvc>=3.0.0
dvc-gdrive>=2.20.0
mlflow>=2.10.0
python-dotenv>=0.19.0
dagshub>=0.5.0
jsonschema>=4.0.0

# Development and testing
pytest>=7.0.0
pytest-cov>=4.0.0
black>=23.0.0
isort>=5.12.0
flake8>=6.0.0
mypy>=1.0.0

# Frontend dependencies are managed separately in the frontend/package.json file
# To install frontend dependencies, run:
# cd frontend && npm install
